--- android/app/src/main/java/com/zzang/chongdae/data/remote/source/OfferingWriteDataSource.kt
@@ -1,7 +1,0 @@
-package com.zzang.chongdae.data.remote.source
-
-import com.zzang.chongdae.data.remote.dto.request.OfferingWriteRequest
-
-interface OfferingWriteDataSource {
-    suspend fun saveOffering(offeringWriteRequest: OfferingWriteRequest): Result<Unit>
-}
