+++ android/app/src/main/java/com/zzang/chongdae/data/remote/source/OfferingsDataSource.kt
@@ -1,5 +1,6 @@
 package com.zzang.chongdae.data.remote.source
 
+import com.zzang.chongdae.data.remote.dto.request.OfferingWriteRequest
 import com.zzang.chongdae.data.remote.dto.response.OfferingsResponse
 
 interface OfferingsDataSource {
@@ -7,4 +8,6 @@ interface OfferingsDataSource {
         lastOfferingId: Long,
         pageSize: Int,
     ): Result<OfferingsResponse>
+
+    suspend fun saveOffering(offeringWriteRequest: OfferingWriteRequest): Result<Unit>
 }
