--- android/app/src/main/java/com/zzang/chongdae/data/remote/source/impl/OfferingWriteDataSourceImpl.kt
@@ -1,21 +1,0 @@
-package com.zzang.chongdae.data.remote.source.impl
-
-import com.zzang.chongdae.data.remote.api.OfferingsApiService
-import com.zzang.chongdae.data.remote.dto.request.OfferingWriteRequest
-import com.zzang.chongdae.data.remote.source.OfferingWriteDataSource
-
-class OfferingWriteDataSourceImpl(
-    private val offeringsApiService: OfferingsApiService,
-) : OfferingWriteDataSource {
-    override suspend fun saveOffering(offeringWriteRequest: OfferingWriteRequest): Result<Unit> {
-        return runCatching {
-            val response = offeringsApiService.postOfferingWrite(offeringWriteRequest)
-            if (response.isSuccessful) {
-                response.body() ?: error("에러 발생: null")
-            } else {
-                error("에러 발생: ${response.code()}")
-//                throw IllegalStateException()
-            }
-        }
-    }
-}
