+++ android/app/src/main/java/com/zzang/chongdae/data/remote/source/impl/OfferingsDataSourceImpl.kt
@@ -1,6 +1,7 @@
 package com.zzang.chongdae.data.remote.source.impl
 
 import com.zzang.chongdae.data.remote.api.OfferingsApiService
+import com.zzang.chongdae.data.remote.dto.request.OfferingWriteRequest
 import com.zzang.chongdae.data.remote.dto.response.OfferingsResponse
 import com.zzang.chongdae.data.remote.source.OfferingsDataSource
 
@@ -14,4 +15,16 @@ class OfferingsDataSourceImpl(
         runCatching {
             service.getArticles(lastOfferingId, pageSize).body() ?: throw IllegalStateException()
         }
+
+    override suspend fun saveOffering(offeringWriteRequest: OfferingWriteRequest): Result<Unit> {
+        return runCatching {
+            val response = service.postOfferingWrite(offeringWriteRequest)
+            if (response.isSuccessful) {
+                response.body() ?: error("에러 발생: null")
+            } else {
+                error("에러 발생: ${response.code()}")
+//                throw IllegalStateException()
+            }
+        }
+    }
 }
