+++ android/app/src/main/java/com/zzang/chongdae/domain/repository/OfferingsRepository.kt
@@ -7,4 +7,19 @@ interface OfferingsRepository {
         lastOfferingId: Long,
         pageSize: Int,
     ): List<Offering>
+
+    suspend fun saveOffering(
+        memberId: Long,
+        title: String,
+        productUrl: String?,
+        thumbnailUrl: String?,
+        totalCount: Int,
+        totalPrice: Int,
+        eachPrice: Int?,
+        meetingAddress: String,
+        meetingAddressDong: String?,
+        meetingAddressDetail: String,
+        deadline: String,
+        description: String,
+    ): Result<Unit>
 }
