+++ android/app/src/main/java/com/zzang/chongdae/presentation/view/write/OfferingWriteFragment.kt
@@ -9,8 +9,8 @@ import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import com.zzang.chongdae.R
 import com.zzang.chongdae.data.remote.api.NetworkManager
-import com.zzang.chongdae.data.remote.source.impl.OfferingWriteDataSourceImpl
-import com.zzang.chongdae.data.repository.remote.OfferingWriteRepositoryImpl
+import com.zzang.chongdae.data.remote.source.impl.OfferingsDataSourceImpl
+import com.zzang.chongdae.data.repository.remote.OfferingsRepositoryImpl
 import com.zzang.chongdae.databinding.FragmentOfferingWriteBinding
 import com.zzang.chongdae.presentation.view.MainActivity
 
@@ -21,8 +21,8 @@ class OfferingWriteFragment : Fragment() {
 
     private val viewModel: OfferingWriteViewModel by viewModels {
         OfferingWriteViewModel.getFactory(
-            OfferingWriteRepositoryImpl(
-                OfferingWriteDataSourceImpl(
+            OfferingsRepositoryImpl(
+                OfferingsDataSourceImpl(
                     NetworkManager.offeringsService(),
                 ),
             ),
